+++ R/colourInput.R
@@ -37,6 +37,7 @@
 #' of a HEX value when possible.
 #' @param closeOnClick If \code{TRUE}, then the colour selection panel will close
 #' immediately after selecting a colour.
+#' @param width The width of the input, e.g. `"400px"` or `"100%"`
 #' @seealso \code{\link[colourpicker]{updateColourInput}}
 #' \code{\link[colourpicker]{colourPicker}}
 #' @examples
@@ -90,7 +91,8 @@ colourInput <- function(inputId, label, value = "white",
                         showColour = c("both", "text", "background"),
                         palette = c("square", "limited"),
                         allowedCols = NULL, allowTransparent = FALSE,
-                        returnName = FALSE, closeOnClick = FALSE) {
+                        returnName = FALSE, closeOnClick = FALSE,
+                        width = NULL) {
   # sanitize the arguments
   showColour <- match.arg(showColour)
   palette <- match.arg(palette)
@@ -98,16 +100,19 @@ colourInput <- function(inputId, label, value = "white",
   value <- restoreInput(id = inputId, default = value)
 
   # declare dependencies
-  shiny::addResourcePath("colourpicker-binding",
-                         system.file("srcjs", package = "colourpicker"))
-  shiny::addResourcePath("colourpicker-lib",
-                         system.file("www", "shared", "colourpicker", package = "colourpicker"))
   deps <- list(
     htmltools::htmlDependency(
-      "colourpicker-binding", "0.1.0", c(href = "colourpicker-binding"),
-      script = "input_binding_colour.js"),
+      name = "colourpicker-binding",
+      version = as.character(utils::packageVersion("colourpicker")),
+      package = "colourpicker",
+      src = "srcjs",
+      script = "input_binding_colour.js"
+    ),
     htmltools::htmlDependency(
-      "colourpicker-lib", "0.1.0", c(href = "colourpicker-lib"),
+      name = "colourpicker-lib",
+      version = "1.6",
+      package = "colourpicker",
+      src = "www/shared/colourpicker",
       script = "js/colourpicker.min.js",
       stylesheet = "css/colourpicker.min.css"
     )
@@ -147,6 +152,7 @@ colourInput <- function(inputId, label, value = "white",
   inputTag <-
     shiny::div(
       class = "form-group shiny-input-container",
+      style = htmltools::css(width = htmltools::validateCssUnit(width)),
       `data-shiny-input-type` = "colour",
       label %AND% shiny::tags$label(label, class = "control-label", `for` = inputId),
       inputTag