+++ NEWS.md
@@ -1,3 +1,8 @@
+# colourpicker 1.3.0 (2023-08-19)
+
+- Add `width` parameter to `colourInput()` (#52)
+- Don't use `addResourcePath()` and instead use `htmlDependency()` (#43)
+
 # colourpicker 1.2.0 (2022-10-28)
 
 - Add CSS class to input label (#48)